refactor: reorganize project structure and extract interfaces
- Move Firebase configuration from `lib/firebase.ts` to `lib/configs/firebase.ts`
- Extract TypeScript interfaces from `mockData.ts`, `useReports.ts`, and `useSettings.ts` into a dedicated `lib/interfaces/` directory
- Update imports across the application to reflect new file locations
- Improve variable naming (e.g., `e` to `event`, `e` to `error`) for better clarity
-
Fix minor CSS overflow issue in `globals.css`
- Refine UI styles in `BottomNav` and remove unused imports/logs

:wq
