+++ app/globals.css
@@ -196,6 +196,7 @@
 
 html {
   scroll-behavior: smooth;
+  overflow-y: scroll;
 }
 
 body {
