+++ lib/configs/firebase.ts
