+++ lib/interfaces/PerCatNotificationPref.ts
@@ -1,0 +1,6 @@
+export interface PerCatNotificationPref {
+  catId: string;
+  catName: string;
+  healthAlerts: boolean;
+  visitAlerts: boolean;
+}
