+++ lib/interfaces/ReportConfig.ts
@@ -1,0 +1,6 @@
+export interface ReportConfig {
+  catId: string | "all";
+  dateRange: "7" | "30" | "90" | "custom";
+  customStartDate?: string;
+  customEndDate?: string;
+}
